新增订单信息功能

+++ lams/lamsserver/lams-mapper/src/main/java/com/bupt/lams/mapper/OrderAssetMapper.java
@@ -17,5 +17,5 @@ public interface OrderAssetMapper {
      */
     Long getLatestOidByAid(Long aid);
 
-    void deleteManyByOids(List<Integer> oids);
+    void deleteManyByOids(List<Long> oids);
 }
